fix: docker env not clear log

+++ packages/app/server/index.ts
@@ -19,15 +19,20 @@ const app = new Koa()
 const port = process.env.PORT || 9090
 
 // 每7天清理一下log
-const cleanLogTask = nodeSchedule.scheduleJob('0 0 * * 0', async () => {
-  if (await checkPathExist(paths.logFile)) {
-    fs.rm(paths.logFile)
-  }
-})
+let cleanLogTask: nodeSchedule.Job
+if (process.env.DOCKER !== 'true') {
+  cleanLogTask = nodeSchedule.scheduleJob('0 0 * * 0', async () => {
+    if (await checkPathExist(paths.logFile)) {
+      fs.rm(paths.logFile)
+    }
+  })
+}
 
 exitHook(() => {
   cancelAllSchedule()
-  cleanLogTask.cancel()
+  if (cleanLogTask) {
+    cleanLogTask.cancel()
+  }
 })
 
 app
